+++ test/resampling/uncertain_datasets/test_interpolation.jl
@@ -4,10 +4,10 @@ y = rand(10)
 xgrid = 0:0.5:10 
 
 # Test that interpolation functions work
-y_intp = linear_interpolation(x, y, xgrid)
+y_intp = linear_interpolation(x, y)
 
-@test length(y_intp) == length(xgrid)
-@test y_intp isa Vector{Float64}
+@test length(y_intp(xgrid)) == length(xgrid)
+@test y_intp(xgrid) isa Vector{Float64}
 
 # Test that interpolation work for uncertain index-value datasets
 timeinds = [UncertainValue(Normal, i, 0.1) for i = 1:5]
@@ -18,3 +18,6 @@ grid = RegularGrid(0, 5, 0.4)
 
 @test length(resample(d, grid)) == 2
 @test length(resample(d, StrictlyIncreasing(), grid)) == 2
+
+
+@test create_interp_scheme(1:10, rand(10), RegularGrid(0, 1, 0.1)) isa InterpolationScheme1D